새로운 Operator 클래스를 사용하도록 수정
